Moved service logic to dao layer
